[Chore] #23 - 코드 정리
